create components for top_panel and central_panel with their own .rs file
