+++ src/lib.rs
@@ -2,5 +2,7 @@
 
 mod app;
 pub use app::TemplateApp;
+mod central_panel;
 mod rustbook_code_blocks;
 mod side_panel;
+mod top_panel;
