+++ src/top_panel.rs
@@ -1,0 +1,22 @@
+use eframe::egui;
+
+pub fn top_panel_ui(ctx: &egui::Context) {
+    egui::TopBottomPanel::top("top_panel").show(ctx, |ui| {
+        // The top panel is often a good place for a menu bar:
+
+        egui::menu::bar(ui, |ui| {
+            // NOTE: no File->Quit on web pages!
+            let is_web = cfg!(target_arch = "wasm32");
+            if !is_web {
+                ui.menu_button("File", |ui| {
+                    if ui.button("Quit").clicked() {
+                        ctx.send_viewport_cmd(egui::ViewportCommand::Close);
+                    }
+                });
+                ui.add_space(16.0);
+            }
+
+            egui::widgets::global_theme_preference_buttons(ui);
+        });
+    });
+}
